+++ src/main/resources/neo4j.platform
@@ -1,1 +1,1 @@
-nl.tudelft.graphalytics.neo4j.Neo4jPlatform
+science.atlarge.graphalytics.neo4j.Neo4jPlatform
